+++ NEWS.md
@@ -1,5 +1,7 @@
 # pct (development version)
 
+- Enables download of other OD datasets with `get_od()` (#66)
+
 # pct 0.5.0
 
 - New `uptake_pct_govtarget_2020()` and `uptake_pct_godutch_2020()` functions approximate the uptake models used in the updated 2020 release of the PCT results